general refactoring

+++ src/main/java/com/musala/delivery/drones/entities/dto/DroneDto.java
@@ -17,5 +17,5 @@ public class DroneDto {
 
 	private EStatus state;
 
-	private Float weightLimit;
+	private Double weightLimit;
 }
