+++ src/main/java/com/musala/delivery/drones/entities/dto/MedicationRequestDto.java
@@ -10,7 +10,7 @@ public class MedicationRequestDto {
 
 	private String name;
 
-	private float weight;
+	private Double weight;
 
 	private String image;
 }
